Updates to improve sandbox DX flow and mobile usage
