+++ docs/composedb/getting-started.mdx
@@ -13,12 +13,6 @@ In this getting started guide, you will learn how to:
 3. [Interact with data](./interact-with-data.mdx)
 -->
 
-### But First - ComposeDB API Sandbox
-Before moving on to setting up your local environment, creating composites, and querying your local node, test your GraphQL query skill by executing requests to a live ComposeDB endpoint directly from your browser! 
-
-### [ComposeDB Sandbox](/docs/composedb/sandbox) 
-<br />
-
 Follow the steps below to get started building a social application on ComposeDB in your local environment.
 
 <DocCardList />
