+++ package.json
@@ -36,6 +36,7 @@
     "graphql-ws": "^5.14.2",
     "prism-react-renderer": "^2.1.0",
     "react": "^18.2.0",
+    "react-device-detect": "^2.2.3",
     "react-dom": "^18.2.0",
     "react-ga4": "^2.1.0",
     "react-scripts": "5.0.1",
